Migrate zarr tests
